🚚 Move animation runners to separate file
