+++ src/index.ts
@@ -7,15 +7,12 @@ export {
   atan, atan2, min, max, toRad, toDeg,
 } from "./Math";
 export { default as interpolateColors } from "./Colors";
+export { runSpring, runTiming } from './AnimationRunners'
 
 const {
   event,
-  spring,
   cond,
-  set,
   clockRunning,
-  startClock,
-  stopClock,
   Value,
   add,
   multiply,
@@ -24,8 +21,6 @@ const {
   divide,
   sub,
   eq,
-  block,
-  debug,
   timing,
   Clock,
   Node,
@@ -61,61 +56,6 @@ export const lookup = (
   index: Adaptable<number>,
   notFound: Node = new Value(),
 ) => array.reduce((acc, v, i) => cond(eq(i, index), v, acc), notFound);
-
-export function runSpring(clock: Clock, value: Adaptable<number>, dest: Adaptable<number>) {
-  const state = {
-    finished: new Value(0),
-    velocity: new Value(0),
-    position: new Value(0),
-    time: new Value(0),
-  };
-
-  const config = {
-    toValue: new Value(0),
-    damping: 7,
-    mass: 1,
-    stiffness: 121.6,
-    overshootClamping: false,
-    restSpeedThreshold: 0.001,
-    restDisplacementThreshold: 0.001,
-  };
-
-  return block([
-    cond(clockRunning(clock), 0, [
-      set(state.finished, 0),
-      set(state.time, 0),
-      set(state.position, value),
-      set(state.velocity, 0),
-      set(config.toValue, dest),
-      startClock(clock),
-    ]),
-    spring(clock, state, config),
-    cond(state.finished, debug("stop clock", stopClock(clock))),
-    state.position,
-  ]);
-}
-
-export function runTiming(clock: Clock, value: Adaptable<any>, config: TimingConfig) {
-  const state = {
-    finished: new Value(0),
-    position: new Value(0),
-    time: new Value(0),
-    frameTime: new Value(0),
-  };
-
-  return block([
-    cond(clockRunning(clock), 0, [
-      set(state.finished, 0),
-      set(state.time, 0),
-      set(state.position, value),
-      set(state.frameTime, 0),
-      startClock(clock),
-    ]),
-    timing(clock, state, config),
-    cond(state.finished, debug("stop clock", stopClock(clock))),
-    state.position,
-  ]);
-}
 
 // ## Transformations
 export const translateZ = (perspective: Adaptable<number>, z: Adaptable<number>) => (
